# api.yml
openapi: '3.0.3'
info:
  title: Url Shortener Api
  version: '1.0'
servers:
  - url: https://go-project-278-migu.onrender.com/api
  - url: http://dev.home:8080/api
paths:
  /links:
    get:
      summary: List of links
      description: Returns list of links
      operationId: GetLinks
      parameters:
        - name: range
          in: query
          required: false
          description: Range in format [start, end]
          schema:
            type: string
            example: "[0, 10]"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkList"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: New link
      description: Creates new link
      operationId: CreateLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkParams"
      responses:
        '204':
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /links/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
    get:
      summary: Link by id
      description: Returns link by id
      operationId: GetLink
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update link by id
      description: Updates link by id
      operationId: UpdateLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkParams"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Remove link by id
      description: Removes link by id
      operationId: RemoveLink
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /link_visits:
    get:
      summary: List of link visits
      description: Returns list of link visits
      operationId: GetLinkVisits
      parameters:
        - name: range
          in: query
          required: false
          description: Range in format [start, end]
          schema:
            type: string
            example: "[0, 10]"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkVisitList"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
security:
  - defaultApiKey: []
components:
  schemas:
    LinkList:
      type: array
      items:
        $ref: "#/components/schemas/Link"
    Link:
      type: object
      required:
        - id
        - original_url
        - short_name
        - short_url
      properties:
        id:
          type: integer
          description: Link id
          example: 1
        original_url:
          type: string
          description: Link url
          example: "https://google.com"
        short_name:
          type: string
          description: Link short url code
          example: "ABC123"
        short_url:
          type: string
          description: Link short url
          example: "{server}/r/ABC123"
    LinkParams:
      type: object
      required:
        - original_url
      properties:
        original_url:
          type: string
          description: Link url
          example: "https://google.com"
        short_name:
          type: string
          description: Link code
          example: "ABC123"
          maxLength: 50
          minLength: 6
    LinkVisitList:
      type: array
      items:
        $ref: "#/components/schemas/LinkVisit"
    LinkVisit:
      type: object
      required:
        - id
        - link_id
        - ip
        - user_agent
        - status
        - created_at
      properties:
        id:
          type: integer
          description: Link visit id
          example: 1
        link_id:
          type: integer
          description: Link id
          example: 1
        ip:
          type: string
          description: Visiter ip
          example: 10.0.0.1
        user_agent:
          type: string
          description: Visiter user agent
          example: Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:124.0) Gecko/20100101 Firefox/124.0
        status:
          type: integer
          description: Request status
          example: 302
        created_at:
          type: string
          description: Visit create time
          example: ""
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error name
        message:
          type: string
          description: Error message
  securitySchemes:
    defaultApiKey:
      type: apiKey
      name: api-key
      in: header