# db/generated/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"database/sql"
	"time"
)

type Link struct {
	ID          int64
	OriginalUrl string
	ShortName   string
	CreatedAt   time.Time
	UpdatedAt   time.Time
}

type Visit struct {
	ID        int64
	LinkID    int64
	Ip        sql.NullString
	UserAgent sql.NullString
	Referer   sql.NullString
	Status    int16
	CreatedAt time.Time
}